moved Nomination components and custom hook to (nomination) dir
